Move Advisor to optimisers module
