+++ test/unit/pipelines/test_pipeline_node_factory.py
@@ -1,6 +1,6 @@
 import pytest
 
-from fedot.core.composer.advisor import PipelineChangeAdvisor
+from fedot.core.optimisers.advisor import PipelineChangeAdvisor
 from fedot.core.optimisers.gp_comp.pipeline_composer_requirements import PipelineComposerRequirements
 from fedot.core.optimisers.graph import OptNode
 from fedot.core.pipelines.pipeline_node_factory import PipelineOptNodeFactory
